fix(keys-to-objects): keysToObjects function object merge
Previoulsy used Object.assign to combine the two objects: fileTree & fileTreeProps, which caused
fileTreeProps to overwrite fileTree, which was not what I wanted. Wrote a function that will
recursively check b's properties, if the property is not an object and does not exist on `a`, then
create that property on `a`. If the prop is contained in `b` and an object we run the function again
passing the same property of both objects.